refactor out SimpleTimedSqlConnectionSupplier
